+++ Sources/UIApplicationDelegate.swift
@@ -6,32 +6,32 @@
 //  Copyright © 2018 flowkey. All rights reserved.
 //
 
-/// Not currently implemented for crossplatform UIKit
-public enum UIApplicationLaunchOptionsKey: String {
-    case dummy
+public struct UIApplicationLaunchOptionsKey: RawRepresentable, Hashable {
+    public var rawValue: String
+    public init(rawValue: String) {
+        self.rawValue = rawValue
+    }
 }
-
 
 public protocol UIApplicationDelegate: class {
     init()
-    var window: UIWindow? { get set } // Not sure what's supposed to happen when you set this
+    var window: UIWindow? { get set }
 
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplicationLaunchOptionsKey: Any]?) -> Bool
     func applicationWillTerminate(_ application: UIApplication)
 
-    // NOTE: THE FOLLOWING ARE NOT YET IMPLEMENTED!
-    func applicationDidEnterBackground(_ application: UIApplication)
     func applicationWillEnterForeground(_ application: UIApplication)
-
     func applicationDidBecomeActive(_ application: UIApplication)
+
     func applicationWillResignActive(_ application: UIApplication)
+    func applicationDidEnterBackground(_ application: UIApplication)
 }
 
 // Swift doesn't have optional protocol requirements like objc does, so provide defaults:
 public extension UIApplicationDelegate {
     func applicationWillTerminate(_ application: UIApplication) {}
-    func applicationDidEnterBackground(_ application: UIApplication) {}
     func applicationWillEnterForeground(_ application: UIApplication) {}
     func applicationDidBecomeActive(_ application: UIApplication) {}
     func applicationWillResignActive(_ application: UIApplication) {}
+    func applicationDidEnterBackground(_ application: UIApplication) {}
 }
